+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added overload to GuidUtils that takes an enumerable of guids and combines them into a new deterministic guid
 
 ### Changed
  - A SafeTimer constructor that executes the callback immediately now does this instead of waiting infinitely